requirements.txt
fastapi
uvicorn
pymongo
python-dateutil
python-dotenv
